Data validations and translation
